requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlmodel==0.0.14

# Celery and Redis
celery==5.3.4
redis==5.0.1

# PDF processing
PyMuPDF==1.23.8
pdfminer.six==20221105
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0

# AI/ML
openai>=1.12.0

# Utilities
python-dotenv==1.0.0

